Add files via upload
Here is the SQL text for a project I did using Joins for Exploratory Data Analysis on old 2014 data. Along with it are modified excel documents from which I extracted the datasets into SQL. The links to each original, unaltered source are in comments within the SQL document for anyone interested!